+++ examples/data-sources/nxos_pim_static_rp/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_pim_static_rp" "example" {
+  vrf_name = "default"
+  address  = "1.2.3.4"
+}