+++ examples/resources/nxos_pim_static_rp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_static_rp.example "sys/pim/inst/dom-[default]/staticrp/rp-[1.2.3.4]"