+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Add nxos_pim_vrf resource and data source
 - Add nxos_pim_interface resource and data source
 - Add nxos_pim_static_rp_policy resource and data source
+- Add nxos_pim_static_rp resource and data source
 
 ## 0.2.3